@@ -14,10 +14,11 @@
     <packaging>jar</packaging>
 
     <properties>
-        <java.version>8</java.version>
+        <java.version>17</java.version>
         <tapdata.pdk.api.version>2.0.0-SNAPSHOT</tapdata.pdk.api.version>
         <kafka.version>2.8.2</kafka.version>
         <slf4j.version>1.7.25</slf4j.version>
+        <caffeine.version>3.2.0</caffeine.version>
     </properties>
 
     <dependencies>
@@ -61,6 +62,11 @@
             <scope>test</scope>
         </dependency>
 
+        <dependency>
+            <groupId>com.github.ben-manes.caffeine</groupId>
+            <artifactId>caffeine</artifactId>
+            <version>${caffeine.version}</version>
+        </dependency>
 
         <dependency>
             <groupId>org.apache.logging.log4j</groupId>
@@ -103,6 +109,11 @@
             <groupId>org.mockito</groupId>
             <artifactId>mockito-inline</artifactId>
             <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>io.tapdata</groupId>
+            <artifactId>pdk-error-code</artifactId>
+            <scope>provided</scope>
         </dependency>
     </dependencies>
 
