@@ -75,7 +75,16 @@ public class GreenplumConnector extends PostgresConnector {
             return "null";
         });
         codecRegistry.registerFromTapValue(TapArrayValue.class, "text", tapValue -> {
-            if (tapValue != null && tapValue.getValue() != null) return toJson(tapValue.getValue());
+            if (tapValue != null && tapValue.getValue() != null) {
+                if (tapValue.getOriginType().endsWith(" array")) {
+                    if (tapValue.getOriginValue() instanceof PgArray) {
+                        return tapValue.getOriginValue();
+                    } else {
+                        return tapValue.getValue();
+                    }
+                }
+                return toJson(tapValue.getValue());
+            }
             return "null";
         });
 
@@ -185,6 +194,10 @@ public class GreenplumConnector extends PostgresConnector {
             greenplumTest.testOneByOne();
             return connectionOptions;
         }
+    }
+
+    public List<TapTable> discoverPartitionInfo(List<TapTable> tapTableList) {
+        return tapTableList;
     }
 
 }
