@@ -534,7 +534,7 @@ public class MysqlConnector extends CommonDbConnector {
         }
     }
 
-    private void writeRecord(TapConnectorContext tapConnectorContext, List<TapRecordEvent> tapRecordEvents, TapTable tapTable, Consumer<WriteListResult<TapRecordEvent>> consumer) throws Throwable {
+    protected void writeRecord(TapConnectorContext tapConnectorContext, List<TapRecordEvent> tapRecordEvents, TapTable tapTable, Consumer<WriteListResult<TapRecordEvent>> consumer) throws Throwable {
         beforeWriteRecord(tapTable);
         List<String> autoIncFields = writtenTableMap.get(tapTable.getId()).getValue(HAS_AUTO_INCR, new ArrayList<>());
         String insertDmlPolicy = tapConnectorContext.getConnectorCapabilities().getCapabilityAlternative(ConnectionOptions.DML_INSERT_POLICY);
@@ -793,7 +793,7 @@ public class MysqlConnector extends CommonDbConnector {
     protected void queryByAdvanceFilterWithOffset(TapConnectorContext connectorContext, TapAdvanceFilter filter, TapTable table, Consumer<FilterResults> consumer) throws Throwable {
         String sql = commonSqlMaker.buildSelectClause(table, filter, false) + getSchemaAndTable(table.getId()) + commonSqlMaker.buildSqlByAdvanceFilter(filter);
         int batchSize = null != filter.getBatchSize() && filter.getBatchSize().compareTo(0) > 0 ? filter.getBatchSize() : BATCH_ADVANCE_READ_LIMIT;
-        jdbcContext.query(sql, resultSet -> {
+        mysqlJdbcContext.queryWithStream(sql, resultSet -> {
             FilterResults filterResults = new FilterResults();
             //get all column names
             Set<String> dateTypeSet = dateFields(table);
@@ -812,7 +812,7 @@ public class MysqlConnector extends CommonDbConnector {
         });
     }
 
-    private Set<String> dateFields(TapTable tapTable) {
+    protected Set<String> dateFields(TapTable tapTable) {
         Set<String> dateTypeSet = new HashSet<>();
         tapTable.getNameFieldMap().forEach((n, v) -> {
             switch (v.getTapType().getType()) {
@@ -1017,17 +1017,32 @@ public class MysqlConnector extends CommonDbConnector {
             sb.append("unique ");
         }
         sb.append("index ");
-        if (EmptyKit.isNotBlank(tapIndex.getName())) {
-            sb.append(escapeChar).append(tapIndex.getName()).append(escapeChar);
-        } else {
-            String indexName = DbKit.buildIndexName(tapTable.getId());
-            tapIndex.setName(indexName);
-            sb.append(escapeChar).append(indexName).append(escapeChar);
-        }
+        String indexName = DbKit.buildIndexName(tapTable.getId(), tapIndex, commonDbConfig.getMaxIndexNameLength());
+        tapIndex.setName(indexName);
+        sb.append(escapeChar).append(indexName).append(escapeChar);
         sb.append(" on ").append(getSchemaAndTable(tapTable.getId())).append('(')
                 .append(tapIndex.getIndexFields().stream().map(f -> escapeChar + f.getName() + escapeChar + (EmptyKit.isNotNull(f.getSubPosition()) ? "(" + f.getSubPosition() + ")" : " ") + (f.getFieldAsc() ? "asc" : "desc"))
                         .collect(Collectors.joining(","))).append(')');
         return sb.toString();
+    }
+
+    protected List<String> getAfterUniqueAutoIncrementFields(TapTable tapTable, List<TapIndex> indexList) {
+        if (!mysqlConfig.getCreateAutoInc()) return new ArrayList<>();
+        String sql = "ALTER TABLE `%s` MODIFY COLUMN `%s` %s AUTO_INCREMENT";
+        List<String> uniqueFields = new ArrayList<>();
+        List<String> uniqueAutoIncrementFields = new ArrayList<>();
+        indexList.forEach(index -> {
+            if (index.isUnique()) {
+                uniqueFields.addAll(index.getIndexFields().stream().map(TapIndexField::getName).collect(Collectors.toList()));
+            }
+        });
+        tapTable.getNameFieldMap().values().forEach(f -> {
+            if (f.getAutoInc() && !f.getPrimaryKey() && uniqueFields.contains(f.getName())) {
+                uniqueAutoIncrementFields.add(f.getName());
+            }
+        });
+
+        return uniqueAutoIncrementFields.stream().map(f -> String.format(sql, tapTable.getId(), f, tapTable.getNameFieldMap().get(f).getDataType())).collect(Collectors.toList());
     }
 
 }
