@@ -16,7 +16,7 @@ import static io.tapdata.base.ConnectorBase.testItem;
 public class OceanbaseTest extends MysqlConnectionTest implements AutoCloseable {
 
     public OceanbaseTest(OceanbaseConfig oceanbaseConfig, Consumer<TestItem> consumer, ConnectionOptions connectionOptions) {
-        super(oceanbaseConfig, consumer,connectionOptions);
+        super(oceanbaseConfig, consumer, connectionOptions);
         if (!ConnectionTypeEnum.TARGET.getType().equals(commonDbConfig.get__connectionType())) {
             testFunctionMap.put("testObLogProxy", this::testObLogProxy);
         }
@@ -34,13 +34,14 @@ public class OceanbaseTest extends MysqlConnectionTest implements AutoCloseable 
 
     public Boolean testObLogProxy() {
         try {
-            NetUtil.validateHostPortWithSocket(commonDbConfig.getHost(), ((OceanbaseConfig) commonDbConfig).getLogProxyPort());
+            NetUtil.validateHostPortWithSocket(((OceanbaseConfig) commonDbConfig).getLogProxyHost(), ((OceanbaseConfig) commonDbConfig).getLogProxyPort());
             consumer.accept(testItem(ITEM_OB_LOG_PROXY, TestItem.RESULT_SUCCESSFULLY, "ObLogProxy is available"));
         } catch (Exception e) {
             consumer.accept(testItem(ITEM_OB_LOG_PROXY, TestItem.RESULT_SUCCESSFULLY_WITH_WARN, "ObLogProxy is not available"));
         }
         return true;
     }
+
     @Override
     protected Boolean testDatasourceInstanceInfo() {
         buildDatasourceInstanceInfo(connectionOptions);
